apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic==2.9.2
aiofiles==24.1.0
cryptography==42.0.8
